libm: default on
Update all `Makefile`s to link `libm` (`-lm`) by default.
Only `icc`, `cce` and `msvc` do not need to link this since they come
with their own, auto-on optimized math libraries.

Just an idea to switch this on, since GCC seams to be the default
compiler in the `Makefile`s. A bit more logic is needed if one wants
to disable this in the other cases.